+++ tuxemon/world/menu_flags.py
